backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.13
botocore==1.42.13
click==8.3.1
environs==14.5.0
fastapi==0.125.0
fastapi_cors==0.0.6
h11==0.16.0
httptools==0.7.1
idna==3.11
jmespath==1.0.1
marshmallow==4.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
